kisaansetu-backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
httpx>=0.25.0
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
Pillow>=10.0.1
geojson>=3.1.0
motor>=3.3.1
pymongo>=4.5.0
anthropic>=0.7.0
certifi>=2023.7.22

chromadb==0.5.3
sentence-transformers==3.0.1
anthropic==0.34.0
langchain-core==0.3.0
